- Made ArtistRequest - Made ErrorResponse model - Added ExceptionHandlingMiddleware with errors handling (NotFoundException, BadRequestException, IncorrectOldPasswordException, AlreadyAddedException) - Added PasswordChange model - Made UserLogin - Made error handling in requests models - Made fixes and improvements.

+++ HomeLib/HomeLib.Services/Services/UserService.cs
@@ -3,7 +3,7 @@ using HomeLib.Core.Interfaces;
 using HomeLib.Core.Interfaces.ForRepositories;
 using HomeLib.Core.Model;
 using HomeLib.Infrastructure;
-using HomeLib.Infrastructure.Repositories;
+using HomeLib.Core.Exceptions;
 using Microsoft.AspNetCore.Identity;
 
 namespace HomeLib.Services.Services;
@@ -15,13 +15,18 @@ public class UserService(IUsersRepository usersRepository, JwtService jwtService
         return await usersRepository.GetAllUsersAsync();
     }
 
-    public async Task<User?> GetUsersById(Guid id)
+    public async Task<User?> GetUserById(Guid id)
     {
-        return await usersRepository.GetUsersByIdAsync(id);
+        if (id == Guid.Empty) throw new BadRequestException("Invalid Id");
+        var user = await usersRepository.GetUserByIdAsync(id);
+        return user ?? throw new NotFoundException($"User with {id} id not found");
     }
 
-    public async Task RegisterUser(string login, string password, string name)
+    public async Task<Guid> RegisterUser(string login, string password, string name)
     {
+        var existingUser = await usersRepository.GetUserByLoginAsync(login);
+        if (existingUser != null) throw new AlreadyAddedException($"User already with {login} login exists");
+
         var user = new User
         {
             Name = name,
@@ -30,26 +35,42 @@ public class UserService(IUsersRepository usersRepository, JwtService jwtService
         };
 
         var hashPassword = new PasswordHasher<User>().HashPassword(user, password);
+
         user.Password = hashPassword;
         await usersRepository.AddUserAsync(user);
-    }
-
-    public async Task<User?> GetUserByLogin(string login)
-    {
-        return (await usersRepository.GetUsersByLoginAsync(login));
+        return user.Id;
     }
 
     public async Task<string> Login(string login, string password)
     {
-        var account = await usersRepository.GetUsersByLoginAsync(login);
+        var account = await usersRepository.GetUserByLoginAsync(login);
+        if (account == null) throw new BadRequestException("Invalid login or password");
+
         var result = new PasswordHasher<User>().VerifyHashedPassword(account, account.Password, password);
         return result == PasswordVerificationResult.Failed
-            ? throw new ApplicationException("Invalid login or password")
+            ? throw new BadRequestException("Invalid login or password")
             : jwtService.GenerateJwtToken(account);
     }
 
     public async Task DeleteUserAsync(Guid id)
     {
+        if (id == Guid.Empty) throw new BadRequestException("Invalid id");
         await usersRepository.DeleteUserAsync(id);
+    }
+
+    public async Task UpdateUserPassword(Guid id, string newPassword, string oldPassword)
+    {
+        if (id == Guid.Empty) throw new BadRequestException("Invalid id");
+        var user = await usersRepository.GetUserByIdAsync(id);
+        if (user == null) throw new NotFoundException($"User with {id} not found");
+
+        var result = new PasswordHasher<User>().VerifyHashedPassword(user, user.Password, oldPassword);
+        if (result == PasswordVerificationResult.Failed)
+            throw new IncorrectOldPasswordException("Incorrect old password");
+
+        var hashPassword = new PasswordHasher<User>().HashPassword(user, newPassword);
+        user.Password = hashPassword;
+        user.UpdatedAt = DateTime.UtcNow;
+        await usersRepository.SaveChangesAsync();
     }
 }
